feat(pinned): share a bottom sheet for panel details
Opening a data issue from the pinned rail slid a 570px drawer out beside
the rail, so the detail read as a second surface competing with the panel
its list lives in. Show it as a bottom sheet over the panel body instead,
which is what the match panel already did for potential matches.

That made the behaviour a second consumer, so extract it: PinnedDetailSheet
owns the scrim, the slide-up and the placement, and match moves onto it.

The sheet is positioned absolutely rather than teleported. The rail keeps
backgrounded panels alive with <KeepAlive>, and a deactivated panel's
teleported content leaks out of its target, so a teleporting overlay is
not usable inside a panel. That constraint lived only in the match panel's
markup; now a new panel gets it by construction.

Chrome stays with each panel: match keeps its compact title bar, while the
issue detail brings its own header and action row. Only the behaviour is
shared.

The full data-issues page keeps its right-hand drawer, so the detail body
moves to DataIssueDetailContent and both containers render it.
